gui cont tabs: Use ProgressRadioButtons in Cryo Chamber tab and MIMAS alignment tab
This helps to specify all icons at the same place, and keep the tab code
simpler. It also solves the issue that when pressing a button, it would
first turn green, and after confirming the move, it would turn orange.
It now is immediately orange.

Also add very basic support of the MIMAS on the chamber tab. The buttons
don't have any effect, but at least the tab is displayed.