SDEV-1247 fix calibration region overlay appearing when the button was deselected
When selecting and deselecting a region of calibration the coordinates of active but deselected scintillators can update unexpectedly. When the coordinates change all active scintllators are checked and re-drawn based on the coordinates.
Because the coordinates had updated even the deselected boxes would be re-drawn.
This fix makes sure that boxes are only drawn based on whether the calibration region box has been selected or de-selected.