gui cont project: removed TODOs and FIXMEs in fastem code
There where two FIXMEs related to using the calib_prefix variable for distinguising between calibration 2 and 3, however here we need to intialize both calibration 2 and 3 so it does not make sense to use the calib_prefix variable.
There was a TODO related to disabling the project list button when running an acquisition and enabling it during calibration, this is already the case so the TODO is unnecessary.