gui cont fastem_acq: sdev1248 fix gui not opening after starting calibration when no ROC is selected
When there is no ROC selected on any of the scintillators and the calibrate button is clicked the gui would disable all the buttons and it would be impossible to cancel the calibration.
With this fix when the calibrate button is clicked it first checks if any of the scintillators have a region of calibration, if not it updates the calibration state and returns.

It was easily possible to click the calibrate button when quickly selecting and de-selecting a roc-box, because there was a wxlimit_invocation of 1 second before updating the calibration controls,
so it took approximately 1 second to disable the calibration button again after quickly de-selecting. The wxlimit_invocation is now set to 0.1 second to make it harder to do this.