model: if base directory is not available use a temporary one
The BASE_DIRECTORY is used as a fixed known location for the connection between the backend and the clients. However, typically it can only be created by root. So if it's present, and the backend is run as standard user, the backend may fail starting. For standard usage, it's not an issue, but for test cases, that can be an issue.

=> In such case, use a temporary folder instead. 

This fixes failing test cases under class ProxyOfProxyTest(unittest.TestCase).

Also do some minor coding styles improvements.