config: extend mimas simulator and add simulator using the Orsay Physics FIB simulator
Add GIS simulator (we probably could use that approach for a lot of the
missing components).
Also adjust the metadata of the stage.

Also introduce a more fancy simulator: the
simulator provided by Orsay Simulator. The main drawback is that it
requires that software, running in a Windows VM.
So we keep the simple mimas simulator, for the cases this fancy
simulator is not available.