fastem gui: fix for pop-up of calibration regions after de-selecting the scintillator
selecting a calibration region in the acquisition tab and afterwards deselecting the scintillator in the chamber tab, caused the calibration region to pop up randomly in the acquisition tab.
With this fix, when a scintillator is removed from the list of active scintillators it will also toggle the is_selected attribute to False.