driver andorcam2: don't allow a default "device" argument
The user has to pass a device number or serial number to connect to a
camera. Until now, if for some reason the user forgot, instead of
explicitly complaining, it would initiliase the "system" library without
connecting to camera, which provides a very odd empty component. (which
is only helpful in some debugging situations).

=> remove the default value, to force a device argument.

Also fix error on terminate() when instantiating a system device.