bugreporter: always store all the YAML files from /usr/share/odemis
Now that we have YAML files that can include other YAML files, it gets
really tricky to know which files to include. It's also not a big
issue to have a few more microscope files. That actually pretty much
always useful.

=> In addition to the MODEL or MODEL_SELECTOR content, also always
include all the YAML files from /usr/share/odemis/. It's still useful
to include MODEL and MODEL_SELECTOR in case they point to files out
of this directory.