driver orsay: improve state error handling
When a component has an error, it's reported in its .state VA.
However, due to how Exceptions are compared in Python, every time the
state is read, it's concidered a new value, and that ends up being
picked up by the GUI, which shows it again.
=> Detect in the driver that the state hasn't really changed, it's just
still the same error, and so do not update it again.