util comp: fix use of MD_POS_ACTIVE_RANGE in generate_zlevels()
MD_POS_ACTIVE_RANGE is a dict of the format axis -> tuple. However, the
code assumed that it would just be a tuple.

=> Look explicitly for the "z" axis.

Added a test cases explicitly for this case.

Also simplify the code, and fix failures when zrange was a tuple while
it was outside of the actutor range.