driver orsay: hide ion-beam.shift complexity
The actual beam shift has to
be adjusted for each probe current, so that the center of the FoV
corresponds to the same point. Also, every time the probe current is
changed, it's reset. We don't want to pass this trouble to the higher
level code. So the "base" shift for each probe current is stored (and
shared via MD_SCAN_OFFSET_CALIB), and it's hidden in the ion-beam.shift, so that a
shift of N µm always stays, even if the current changes.

The fib_beam.shift still allows to observe/affect the low-level value of
the shift.