bugreporter: set POST request content-type as JSON
We upload JSON data, so it should be marked as JSON. Typically, the
content-type doesn't matter. At least osTicket doesn't care.
However, the HTTP server has new security rules that check that content,
and automatically rejected POST requests which, by default, reported
x-www-form-urlencoded.

So this fixes the issue that some users are facing when reporting a bug
and it indicates Error 400.