util conversion: Add **kwargs to the init of YamlExtraDumper, to fix passing along sort_keys
In 18140fbad328520717036585bdaad625deca7b15 the keyword argument sort_keys was removed for compatibility with ubuntu 18.
However this caused the test case and starting of odemis to fail on ubuntu 20, because the class is called with the argument sort_keys in yaml.dump_all().
This fix adds **kwargs to the input arguments of YamlExtraDumper class and does not pass the keyword arguments along, for compatibility with both ubuntu 18 and 20.