acq millingmng: after moving to a site, reset the beam shift
The beam shift is used to compensate for the drift while milling.
Every time we move to a new place on the sample, we have a "fresh canvas",
so we should reset the beam shift. One of the advantage being that it
reduces the chance of getting out of the limits of the shift during
milling.
In addition, we can use the beam shift to compensate for the
impressision of the stage movement.
Of course, there might still have been a drift between the moment the
feature was selected, and the milling, but for now we don't handle that.