gui model StreamView: fix zooming out completely on an Overview viewport
The FixedOverviewView and FeatureOverviewView allow to zoom out a lot
more than standard views. Up to 1m/px. As the range is stored as an int,
when reaching that range, the mpp would be set to 1. To compute the
.fov, there is some code that just does mpp * width in px. That would
result in an int. However .fov is defined as tuple of floats, only.
So, changing the mpp, changed the FoV, which caused an error.

=> Also allow FoV to be ints. There is nothing wrong with int, they are
real numbers too!