driver orsay: fix behaviour of translation and some test cases
The ion-beam.translation VA didn't behave as the convention. For scales > 1,
the computation was done using the size of the pixel, instead of using
the size of the pixel at scale 1.

Also fix some other test cases.

Also adjust the code for the newest Orsay Server, which now returns
sometimes "Name" instead of "name" in the preset attributes.