scripts cl_acquisition: update to use more standard acquisition method
Complete overall of the script to replace the old synchronization based
on the e-beam scanner trigger to a fully software-based synchronization.
It's slower but a lot more reliable, with any kind of hardware and
software.

Also copy a lot of the latest code from the standard SPARC acquisition
to simplify of make it more reliable.