[feat] Support for partial SPARC acquisition
In case a MDStream acquisition fails, don't discard the data acquired so
far.

=> Change the API, in a similar way as the acqmng API: the Future always
returns both the data acquired and the exception. If the acquisition
went fine, the exception is None.

If an exception occured, the GUI will show the error message, and store
the data in a file named "filename-failed...", to make it clear it
didn't work as expected.