odemisd: Do not require that all the components be listed as children of the microscope
If a component was not in listed as child of the microscope, a warning
would be logged... and no one cares.
These components are not "children" in the standard meaning of child
being a component instantated automatically by another componenent at
init.

Actually, for already a long time, what matters is .alive and .ghost on
the microscope. No code seems to ever care about Microscope.children.

However, this takes a lot of time to update for each microscope file.
So just assume that if a component is in the microscope file, it should
be instantiated.