util spectrum: make get_angle_per_pixel() really able to fall back to recreating the metadata
If MD_THETA_LIST is not present would try to recreate the metadata...
but it would typically generate the wrong shape as we pass data of CAZYX
while the function expected AC.

Also detect that MD_THETA_LIST with only NaNs is wrong, so that the rest
of Odemis can fall-back to show just pixel values.

Also add test cases!