odemisd: log a better error if component name is missing
With the "!extend" syntax it's relatively easy to forget the name of the
component (if using it to insert a whole component).
It would fail in very confusing way. This was noticed by Stefan Sneep.
=> Add a check to more clearly indicate the error.

Note that ideally we'd link back the error to a specific "extend" file,
however it's tricky because technically the YAML is correct, and by the
name we've loaded it, we have discarded the information.

This is an example of the error generated when the component name is absent without the fix:
2023-05-08 11:19:28,415 ERROR   main:97:        When instantiating file ../mic-odm-yaml/muenster-sparc-3056/sparc2-muenster-lwd-polarizer-sim.odm.yaml
Traceback (most recent call last):
  File "/home/piel/development/odemis/src/odemis/odemisd/main.py", line 88, in __init__
    self._instantiator = modelgen.Instantiator(model_file, settings_file, self,
  File "/home/piel/development/odemis/src/odemis/odemisd/modelgen.py", line 263, in __init__
    self._preparate_microscope()
  File "/home/piel/development/odemis/src/odemis/odemisd/modelgen.py", line 287, in _preparate_microscope
    microscopes = [(n, a) for n, a in self.ast.items() if a.get("class") == "Microscope"]
  File "/home/piel/development/odemis/src/odemis/odemisd/modelgen.py", line 287, in <listcomp>
    microscopes = [(n, a) for n, a in self.ast.items() if a.get("class") == "Microscope"]
AttributeError: 'str' object has no attribute 'get'
2023-05-08 11:19:28,416 ERROR   main:780:       Unexpected error at instantiation

This is an example with the fix:
2023-05-08 11:58:21,547 ERROR   main:777:       Error while parsing file ../mic-odm-yaml/muenster-sparc-3056/sparc2-muenster-lwd-polarizer-sim.odm.yaml:
Definition of component "class" should be a dict but got a str

Note that this also adds checks for a few other errors that would cause
odd error messages (eg, if a component name is not a string).

Also fixes the test cases to run from any directory.