driver xt_client: extra check that the stage is really not moving
Sometimes, half-way through a move the stage is reported not moving for a short while (usually,
when one axis finished moving, and another is about to start.

=> be more careful, and double check before really stopping.

That fixes issues where the stage would be still moving, and the
acquisition would already start, or even more dangerous, on the METEOR,
the FLM objective could be engaged, while the SEM stage is still
rotating.

With this extra check, we can also lower the initial delay of 0.5s before ever
checking that the stage started to move. That means that very short
moves will now be shorter in time too.