CI: add self hosted runner workflow config files
Allows to automatically run all the test cases, on both Ubuntu 18.04 and 20.04. This is done by launching them in a VM on our dedicated computer.
We cannot do it on every commit or pull request because running all the test cases takes around 6h (per VM). However, we can easily do it daily. => Do it every evening on the master, so that on the next morning we see what failed.