plugins quickcl: don't allow the user to minimize the window
The plugin window is modal (with good reasons), so it's not possible to
use the standard GUI while the window is opened. However, in Gnome, as
the dialog is maximized (and normally dialogs can never be maximized),
the minimize and maximize buttons are shown. If the user presses the
"minimize" button, the standard window is shown but disabled, which is
quite confusing.

=> Don't maximize at init, just make "very big". This way there is never
a minimize button, and it's much harder to be confused.

Also fix a division by zero when the data is empty and a logarithmic
scale is used.

Also show the "save" button by default.