acq feature: remove milling angle from CryoFeature
The idea of the milling angle was that the user would be able to select
the angle at which to mill. However, it turns out with more experience
that the milling angle is more a feature of the milling step. We now
have it part of the milling settings. Can remove this attribute, which
was never used.