util tmcmconfig: adjust min_power to work even if device reports 0V
The TMCM6110 used to report ~3V, instead of 12V, when it was only connected via USB only.
However, the latest versions report exactly 0V.
For some reasons, we had set the min_power to 0.1V... just for the sake
of not saying 0V. However saying 0V is fine. This allows to work on the
latest versions of the TMCM6110.