install: add other HDF5 mimetype to desktop file
The latest HDFViewer (3.3.0) comes with a "proper" debian package.
In that package HDF5 files are mapped to the mimetype text/x-hdf5,
while we have mimetype application/x-hdf.

Now, both mimetypes are installed. In practice, only one will be used
for a given file. It's not clear which one, but to be sure that Odemis
Viewer is still associated to HDF5 files, report that Odemis accept both.