driver smaract SmarPod: fix move going sometimes to the wrong place
In some very rare cases, usually a move involving many axes, the stage
would rotate even if not requested. It's because the taget position
requested could be corrupted, as Python memory management doesn't take
into account C functions which use pointers. So we need to hold a
reference to the requested target position explicitly in Python.