[feature] new semnidaq driver: control a SEM via the NI-DAQ API
Depends on the "nidaqmx-python" Python wrapper for using the NI-DAQmx API.
For now tested only on a NI PCIe-6361... and also known to not work
yet on older devices like the PCIe-6251.

It supports all features needed for the SECOM and SPARC, with the
exception of the counter detector. Tested succesfully on a real SEM.

It is possible to test it on a computer without an actual card by using
the NI-DAQmx with a simulated DAQ card. Instructions are in the top
comment of semnidaq.py .

Also introduces two microscope files: one to use a real DAQ board, only,
and one to simulate a SPARC using a simulated NI DAQ board.A