gui comp CryoFeatureOverlay: fix crash if a feature changes not in GUI thread
The current feature and the features status are tracked by the
FeatureOverlay, because a redraw is necessary whenever they change.
Until recently, the status of a feature or the addition of a feature
could only be done via the GUI. So it was always happening the GUI
thread.

However, it should never be assumed that changes to the model only
happen in the main GUI thread. For instance, now that we support
milling, the milling process will update the feature status. This
happens in a background thread.
This caused random GUI crashes.

=> Don't redraw directly in the current thread. Just place a request for
redraw.