fastem: change order of calibration 1 (#2446)
Execute the image rotation pre-alignment before the scan rotation and scan amplitude pre-alignments. The latter two calibrations only work if the pattern is roughly aligned with the diagnostic camera x-axis (+/- 5 degrees).
The image rotation pre-alignment is independent of the scan rotation pre-alignment, the descan gain alignment and the scan amplitude alignment, therefore it can be moved in front of those three.