[refactor] gui: remove pubsub dependency
wxpython is dropping wx.pubsub, as it's now a separate project.
Instead of "adding" this dependency, let's just not use pubsub.

We needed wx.pubsub for only one some behaviour of the GUI: when a
stream is removed by pressing the "-" button on a stream panel.
For this case, we now give a callback function to the StreamBar, which
will be callled on the destroy event.
Note that this also solves another minor issue because the stream
controller used to also listen to the DESTROY event. However, wxPython
only passes the event to one listener. If that listener doesn't "skip"
the event, then the other listeners don't receive the event. So the
StreamController actually was sometimes not cleaning up correctly
already.