gui cont tabs SparcAcquisitionTab: also create a "Temporal Intensity" view if time-correlator present
Until now, it was create in case a monochromator is present (to show the
intensity of the monochromator every 1s), or if there is space.
However, the time correlator stream also needs this view. When there is
EK support, that view would not be created (because there is already
enough view). => explicitly create the view.

Also change the name of the view, from "Monochromator" to a more generic
"Temporal Intensity".