[FIX] andorcam2_test.py and simstreakcam_test.py failing because the setUp was overwritten
In 87ec2f53 a setUp was added with a line to ignore a certain warning. Because TestSimStreakCamGenericCam and TestFake inherit from VirtualTestCam, this meant that the VirtualTestCam setUp was overwritten and not executed. Ignoring the warning is now added to the setUpClass, because it does not need to be re-executed for every test case.