[FIX] millmng_test: test_milling_settings_from_yaml updated to actually test milling
There was a sleep of 13 seconds before checking if the future was running, this got removed in commit 9b2c12cb, because after 13 seconds the future was already done running.
It turns out that the future was done running because the milling had failed, because the stage was not in loading position.
This fix moves the stage to loading position and puts back a sleep of 0.1s to be sure the future is running at the check.