[fix] acq calibration: if multiple background data, average should be the same dtype as raw data
If a background data file contains multiple data, now, the average is
returned (it used to return just the first data). However, numpy.mean()
automatically uses float64 when averaging integers. That clashes with
the rest of the background subtraction code which expect the background
data type to be the same as the actual data.

So in such case, the user would get such error message:
"Incompatible encoding of background data float64 with the angular resolved encoding uint16."

=> convert the average value back to the original data type. It's a tiny
bit less precise, but that's typically really not important as the data
has a large range of values. (and it's still better than returning just
the first background data)