[feature] Compensate chromatic aberration of angular spectrum data
The angular spectrum raw data is typically not straight because of
chromatic aberration. That means the angular values actually (slightly)
varies depending the on the wavelength.
The aberration is already measured at alignment, and stored in the metadata.
However, so far the GUI did nothing with this information. The angles
displayed corresponded to the angles at the center wavelength.

This adds support for compensation, by adjusting the data for each
wavelength independently. This is done just once when loading the data
(and every time a background or coefficient compensation file is
loaded).
This also indirectly means that the 2D view angle vs wavelength is now
linear on the angle axis (instead of being some sort of logarithmic
scale). That makes the graph easier to read.