[fix] handle Fit-to-RGB "tint" for line-cut and raw value display
Now that the Fit-to-RGB option is not a separate option, but just a
special tint, some corner-cases where missed.

* Displaying the "raw" value was showing the whole average instead of the
average for the 3 different bands, as we used to do.
* Displaying a line-cut failed. Instead, now we show it as greyscale (it
doesn't make sense to show it in colour, as it's already spread over the
spectrum).

It used to produce errors like:
14:02:15 ERROR log : Updating LineSpectrumProjection image
Traceback (most recent call last):
File "odemis\acq\stream\_projection.py", line 1491, in _updateImage
File "odemis\util\img.py", line 461, in DataArray2RGB
ValueError: too many values to unpack (expected 3)