[fix] Show proper error if component has no class and no creator
If a component is created by another component, but in the microscope
file it's not declared in the "children", then the backend doesn't know
which component is its creator.

Such error used to be handled properly before because most components
always has as "parent" also the Microscope which *weirdly* declared
every component as its child. However, now that we have removed this
weirdness, a component without creator would have no key "parents". This
raised an exception earlier than expected.

=> handle such case by returning an empty parents list. Then the usual
error is displayed properly again.