[fix] Don't round numbers in plot MarkingLine overlay
That overlay allows the user to read precisely the value of a plot.
However, it used to round large number to sometimes just 3 digits.
So numbers like 123456 would be shown "123000". That's good usage
for the axis ticks, but not in the plot where we want as much info as
possible.
In practice this means that for int's the value will now be always fully
shown.
For floats (which is rare), the value will be either rounded to the
closest int, or if its a small number, show at least 4 digits.