[fix] gui util: adjust timing for fix_static_text_clipping()
The text size only adjust to the right size if the widget is already
shown. However, it can take some time... which depends on a lot of
factors. To make it happen as soon as possible, but still be certain it
happens, we do it multiples time, with different delays: 0.1 and 0.5s.

It turns out that sometimes, even waiting 0.5 s was not enough.
More tests showed that 0.1s is almost nevery helpful.

So instead of adjusting after 0.1s and 0.5s, adjust at 0.5s and 1s.