[fix] driver simstreakcam: avoid error when stopping the acquisition
In some cases, when stopping an acquisition, such error would happen:
ERROR   __init__:707:   Failure while calling repeating timer 'SimCam image generator'
Traceback (most recent call last):
  File "/home/piel/development/odemis/src/odemis/util/__init__.py", line 701, in run
    self.callback()
  File "/home/piel/development/odemis/src/odemis/util/weak.py", line 36, in __call__
    return self.f(ins, *arg, **kwargs)
  File "/home/piel/development/odemis/src/odemis/driver/simstreakcam.py", line 314, in _generate
    exp = timer.period
AttributeError: 'NoneType' object has no attribute 'period'

This is not a big deal, as it's in a thread that's supposed to end when
stopping the acquisition. So it'd be restarted automatically on next
acquisition.
Still, it's nicer to not fail => detect such corner case, and just end
the thread nicely.