[refactor] merge SPARC2 mirror alignment into lens alignment
The mirror and lens alignment mode are very similar, with the hardware
in the same position. In practice, most users only use the lens
alignment mode to do everything. The mirror alignment mode offers a few
extra options, which very few users ever notice.

=> Duplicate this options to the lens alignment mode. Now it's
full-featured to do completely the lens and mirror alignment.

It turns out that this actually simplifies the code, as now everything
is on a single stream, instead of 2.
Next step will be to remove the mirror alignment mode completely, but
it's a little bit convoluted as this is the only mode shown on a SPARC
compact.

Also remove some old code that wasn't needed since the special MoI
stream was removed.

Also add a space in Lens1 and Lens2 labels.