[feature] sparc chamber tab: don't keep focus on mirror switching button
After pressing the "PARK MIRROR" button, it typically gets the focus.
Once it's parked, the button becomes "ENGAGE MIRROR".
That means that as long as the user has the mirror parked, pressing
the "space" key will automatically engage the mirror. Really dangerous.

So now, as soon as the button is pressed, we move the focus to the
Cancel button. This way "space" has a much safer behaviour of stopping
the movement (or doing nothing).

It's still possible to give the focus to the PARK MIRROR button, and
press space, but then it when pressing Tab, or using the mouse first to
give it focus. That a lot more explicit.