[perf] StaticSpectrumStream: avoid running the calibration correction multiple times with the same input
At init, the background and coeficient are initialised to None, but
typically the GUI sets them immediately after to the current value,
which is often None too.
In such case, the calibration correction is done up to 3 times in a
row with the same input (and generates the same output).

For some data, like large EK data, the correction can take ~30s. So this
helps opening some files faster.