[fix] technolution ASMDataFlow test cases
Many of the ASMDataFlow test cases where failing because we were not properly waiting for images to be received.
Now instead of sleeping an arbitrary amount of time the code waits until the data is received, if it is not receieved within a specified time the test case times out.