[SDEV-1556] refactor FASTEM GUI controllers
Swap button object as dictionary key and scintillator number as dictionary value.
This pair helps in easily getting the scintillator number when an event on a button click happens

Remove unnecessary subscriptions of coordinates VA
The unnecessary subscriptions of coordinates VA are removed to avoid calling of callback when the coordinates are changed. Used better function naming to create and remove the calibration overlay when ROC button is pressed or active scintillators are changed in the chamber tab.

Introduce "selected" attribute for FastEMROCOverlay to indicate it should look be drawn with emphasis
Now the enable VA gets back its original meaning that indicate it's not possible to interact with it. Useful during calibration or acquisition.

Adjust the log level while setting new ROC coordinates to DEBUG.