[fix] add rotational axes for Thermofisher when switching using grid mode
Also move the rotational axes when switching using grid mode. In theory, it should not be necessary as switching grid should be just a matter of translation in X&Y. However, if the system is in LOADING position, it's currently allowed to move the GRID. In this case it will go to SEM mode, but with potentially wrong rotation.
So, as a workaround, also rotate the axes. This way, the final position is correct.

Probably the right way would just to separate the grid switch from posture switch. It should be illegal to move grid outside of SEM/FM posture.