[feat] technolution driver: Add support for '.' in the allowed charcters of the storage directory
In the update of the ASM v3 API it is now allowed to have a '.' in the storage directory. If a name starts with a '.' this directory is not taking into account in the post-processing.
For the calibrations we need to acquire an image to update acquisition settings, these acquisitions where previously post-processed which takes up a lot of time. This is now not needed anymore.

SDEV-1539