test case test_tint_to_colormap_unexpected_input in /util/test/img_test.py was failing on Ubuntu 18.04. The expected ValueError was not raised in this case.
On Ubuntu 18.04 RGB values in tuples higher than 255 are somehow
accepted or forced in between the range of 0-255.
The root of this reason is not easily traced as the code is
distributed over multple .py files and/or classes.

To make sure the ValueError exception is raised a check is added
to /util/conversion.py rgb_to_frgb method to make sure all the
values in the tuple are within the range of 0-255.