[fix] adjust histogram range of streams with integer data type
The range used to follow two rules that don't have the same goal:
* between 0, and the smallest power of 2 that fits (to match hardware depth)
* between min value and a width power of 2

The second rule means that if the data ever gets a tiny smaller the
range will immediately change. Also it's impossible to move the slider
ever smaller than the smallest value.

Also it didn't handle signed integers properly, because it would always
clip to 0, even if data is lower than 0.

=> Change the logic. The main goal is get a range that fits the data,
and is slightly bigger so that if the next data is slightly different it
still has chances to stay the same. Now use two new rules:
* add 10% margin on each side
* round each side the nearest power of 2
Take whichever is the smallest, and make it at least 256 wide (8 bits).