[fix] gui SPARC2 alignment tab: show mirror control in lens 1 alignment again
This is an error introduced in commit "gui SPARC2 alignment tab:
completly hide panels which cannot be used". It hided too many panels.

Also fix an issue introduced in commit "merge SPARC2 mirror alignment
into lens alignment": now that the stream allows for background
subtraction, the mirror alignment also has the background subtracted. So
it needs to have a way to reset it too. Especially, the image is not the
same as for the lens/mirror alignement.