[fix] util.img.compactHistogram() could return histograms not fitting the range
If the histogram to compact wasn't a multiple of the requested length,
it would extend it to fit. However it would extend it way too much.
It aimed to return an histogram exactly the size requested. That mained
that there could be many bins at the end of the histogram full of zeros,
and especially way beyond the original range. As it didn't return the
updated range, the caller would interpret it as if the histogram is
compressed. That caused odd computation of outliers in the GUI when the
histogram wasn't a multiple of 256.

=> Don't try to return exactly the requested length. Just the length
that ensures that each bin is the same width. That behaviour is fine
because compactHistogram() is used to just have a smaller histogram, but
the exact length (typically 256px) is never really important.