[config] add auto-completion to "odemis" as synonym for "odemis-cli"
The bash auto-completion script already tried to report it also works
for "odemis", but the way the auto-completion works, each command must
have a file named after it.
So make a soft-link from odemis -> odemis-cli so that the
auto-completion also works for odemis. This is the standard way it's
done on Debian for commands with multiple names.