[fix] acq EnzelPostureManager: report the posture based on both the stage and lens positions
So far on the ENZEL the current posture was detected only based on the
stage position. If all the hardware moves fine, and the user only uses
the GUI, that's sufficient. However, in case of issue with "align"
(ie, the lens stage), the position could still be reported as "fine"
although the lens is at the wrong place.

=> Take the position of the lens into account too.