[tweak] log when a hardware failure is reported
It's not clear how well the users see hardware failure reports in the
notifications. For now, it's not even clear whether Odemis didn't report
it, or it did but the user didn't see it. There is a log in the backend,
but maybe something prevents the GUI from detecting the change of state,
or making the notification.
=> Log when the notification is shown