[fix] dataio tiff: don't be too upset if OME metadata has WL_LIST info but the data has no C dimension
This shouldn't happen... but due to a bug in the TIFF exporter, it could
happen that spectrum data would be stored as the T dimension. So the
OME TIFF metadata would have "Channel" information (WL_LIST), but no C
dimension. In such case, it would fail to load completely the data.

Now just discard the WL_LIST metadata.