[fix] util dataio: handle better the odd cases when dropping a dimensions from DataArrayShadow
In the event that there is no good stream to handle the data,
data_to_static_streams() uses a Static2DStream. If the data has more
than 2 dimensions, it just drops the extra ones. However DataArrayShadow
do not directly support access to the data. As that's what the tiff
importer returns, that would always fail with TIFF.

=> In such case, first convert to a DataArray, before dropping the
dimensions.