[fix] Mock the stage in the fastem mocked test cases.
The stage was not mocked because it previously was not called in the test cases. Commit a1baa166 added the use of guessActuatorMoveDuration to estimate the duration of a move.
This checks if the stage object is an actuator. Since it was not mocked at all and it was just None, the following error was raised: ValueError: The component None should be an actuator, but it is not.
Now the stage is also mocked and the check if it is an actuator passes.