[fix] Fix _updateDrange() to compute a histogram of length multiple of 256
The length of the histogram is computed as "min - max + 1".
To ensure the length is a multiple of 256, we can ensure that min is a
multiple of 256, and that "max + 1" is a multiple of 256.
We used to do that... sometimes. In other cases, the code would round
max to 256 (instead of max + 1).
=> Always round max to 256, and then subtract 1. So that max+1 is always
a multiple of 256.

Note that having a histogram length not a multiple of 256 doesn't cause
functional behaviour issues, but it does slow down compactHistogram().
(And it even used to trigger a bug in compactHistogram(), but that's
fixed).