[feat] fastem acquisition: move stage in the scan direction to reduce the xy-offset between fields
Use the scan-stage instead of the stage component to move the stage during acquisition, this way the stage moves along the scan direction and the fields are placed neatly next to each other.
The ROA coordinates are in stage coordinates, therefore are transformed to scan-stage coordinates when calculating the position of the first tile.