[optim] speed up util.img.findOptimalRange() in good cases
If outliers==0 (which is not very common, but does happen from time to
time), we had a shortcut, but it turned out to not save so much time.
Now implement it a better way, so that it really saves some time (~
twice faster than the normal path).

This mainly should help pass the CI tests all the time.