[fix] xtclient remove aperture index as it is not supported by TFS anymore. (#2570)
TFS removed reading and setting the aperture index from their interface. As we do not use it anyway, it is save to remove it.
Also removed the TODO about the beamlet index updating when changing the multibeam, because the beamlet index is not updated when the multibeam mode is changed.