[fix] xtclient correctly raise hardware error with message
In commit c4b4f0723949748a71f6675b8703405fb3017acc initialization errors in the xtclient were adjusted to report them with more precision.
Howver the OSError was except as `except OSError:` not `except OSError as err`, since `err` was not assigned this caused a problem when an OSError was raised.